refactor: rename navigation to navbarMenu and move to header directory
