+++ src/App.tsx
@@ -3,14 +3,14 @@ import { ContactSection } from './components/ContactSection';
 import { ExperienceSection } from './components/ExperienceSection';
 import Footer from './components/footer/Footer';
 import { HeroSection } from './components/HeroSection';
-import { Navigation } from './components/Navigation';
+import NavbarMenu from './components/header/NavbarMenu';
 import { ProjectSection } from './components/ProjectSection';
 import { SkillsSection } from './components/SkillsSection';
 
 function App() {
   return (
     <div className='min-h-screen bg-background text-foreground'>
-      <Navigation />
+      <NavbarMenu />
       <main>
         <HeroSection />
         <AboutSection />
